requirements.txt
PyMySQL==1.1.1
python-dotenv==1.0.1
blinker==1.9.0
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
elastic-transport==8.15.1
elasticsearch==8.16.0
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
PyMySQL==1.1.1
python-dotenv==1.0.1
redis==5.2.0
urllib3==2.2.3
Werkzeug==3.1.3
